+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 All notable changes to FrameLog will be documented in this file.
 
-## [1.5.1-beta] - 2025-10-25
+## [1.6.1-beta] - 2025-10-25
 
 ### Added
+- IsPrinting - checks whether it is possible to print with this minimum log level
+- SetMinimalLogLevel - sets the minimum log level
 
 ### Fixed
-- ReadFile Return
 
 ### Delete